% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-claude.R
\name{claude_r_package}
\alias{claude_r_package}
\title{Claude Code configuration theme for R packages}
\usage{
claude_r_package()
}
\value{
A \code{prefab_theme} object.
}
\description{
Creates a theme that deploys Claude Code agent settings and rules for an
R package.
}
\examples{
claude_r_package()
}
