% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-project.R
\name{r_targets}
\alias{r_targets}
\title{R targets project theme}
\usage{
r_targets()
}
\value{
A \code{prefab_theme} object.
}
\description{
Creates a theme that scaffolds an R targets project with \verb{_targets.R},
\code{packages.R}, \code{README.md}, and \code{.gitignore}.
}
\examples{
r_targets()
}
