% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_code}
\alias{theme_code}
\title{Print R code that reproduces a theme}
\usage{
theme_code(theme)
}
\arguments{
\item{theme}{A \code{prefab_theme} object.}
}
\value{
The generated R code as a single character string (invisibly).
}
\description{
Prints the R code that would reproduce the given theme via \code{\link[=cat]{cat()}}, and
returns the code invisibly as a single character string.
}
