% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-claude.R
\name{claude_r_analysis}
\alias{claude_r_analysis}
\title{Claude Code configuration theme for R analysis projects}
\usage{
claude_r_analysis()
}
\value{
A \code{prefab_theme} object.
}
\description{
Creates a theme that deploys Claude Code agent settings and rules for an
R analysis project.
}
\examples{
claude_r_analysis()
}
