% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-project.R
\name{r_analysis}
\alias{r_analysis}
\title{R analysis project theme}
\usage{
r_analysis()
}
\value{
A \code{prefab_theme} object.
}
\description{
Creates a theme that scaffolds a simple R analysis project with \code{main.R},
\code{README.md}, and \code{.gitignore}.
}
\examples{
r_analysis()
}
